requirements.txt
torch>=2.2.0
torchvision>=0.17.0
Flask>=2.3.3
Pillow>=10.0.0
numpy>=1.24.3
transformers>=4.35.2
boto3>=1.26.0
python-dotenv>=0.19.0
